requirements.txt
fastapi==0.95.2
uvicorn
jinja2
aiofiles
sqlalchemy
python-dotenv
pydantic[email]
passlib[bcrypt]
pytest
requests
python-jose
pre-commit
python-multipart
databases
databases[sqlite]
asyncpg
aiosqlite
seedir
